+++ modules/alloy/src/main/ts/ephox/alloy/positioning/view/SpotInfo.ts
@@ -1,6 +1,7 @@
 import { BubbleInstance } from '../layout/Bubble';
 import { DirectionAdt } from '../layout/Direction';
 import { BoundsRestriction } from '../layout/LayoutBounds';
+import { Placement } from '../layout/Placement';
 
 export interface SpotInfo {
   readonly x: number;
@@ -9,6 +10,8 @@ export interface SpotInfo {
   readonly direction: DirectionAdt;
   readonly label: string;
   readonly restriction: BoundsRestriction;
+  readonly placement: Placement;
+  readonly alwaysFit: boolean;
 }
 
 const nu = (
@@ -16,15 +19,19 @@ const nu = (
   y: number,
   bubble: BubbleInstance,
   direction: DirectionAdt,
-  restriction: BoundsRestriction,
-  label: string
+  placement: Placement,
+  boundsRestriction: BoundsRestriction,
+  labelPrefix: string,
+  alwaysFit: boolean = false
 ): SpotInfo => ({
   x,
   y,
   bubble,
   direction,
-  restriction,
-  label
+  placement,
+  restriction: boundsRestriction,
+  label: `${labelPrefix}-${placement}`,
+  alwaysFit
 });
 
 export {